refactor: move GetCommandToExec to platform build files
