+++ cmd/exec.go
@@ -220,7 +220,6 @@ func executeExec(ctx context.Context, dev *model.Dev, args []string) error {
 			if err != nil {
 				return err
 			}
-			cmd.SysProcAttr.Foreground = true
 
 			return executor.RunCommand(cmd)
 		}
