+++ cmd/up/exec.go
@@ -22,9 +22,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strings"
-	"syscall"
 
 	"github.com/okteto/okteto/cmd/utils"
 	"github.com/okteto/okteto/pkg/cmd/pipeline"
@@ -59,31 +57,6 @@ type HybridExecCtx struct {
 	Name, Namespace string
 	Client          kubernetes.Interface
 	RunOktetoExec   bool
-}
-
-func (he *hybridExecutor) GetCommandToExec(ctx context.Context, cmd []string) (*exec.Cmd, error) {
-	var c *exec.Cmd
-	if runtime.GOOS != "windows" {
-		c = exec.Command("bash", "-c", strings.Join(cmd, " "))
-	} else {
-		binary, err := expandExecutableInCurrentDirectory(cmd[0], he.workdir)
-		if err != nil {
-			return nil, err
-		}
-		c = exec.Command(binary, cmd[1:]...)
-	}
-
-	c.Env = he.envs
-
-	c.Stdin = os.Stdin
-	c.Stdout = os.Stdout
-	c.Stderr = os.Stderr
-
-	c.Dir = he.workdir
-
-	c.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-
-	return c, nil
 }
 
 func (he *hybridExecutor) RunCommand(cmd *exec.Cmd) error {
